Sửa một số lỗi

+++ app/src/main/java/esolutions/com/gcs_svn_old/com/camera/utility/AsyncCallWS.java
@@ -1186,7 +1186,7 @@ public class AsyncCallWS {
 
                     offset += ChunkSize;
 
-                    int count = (int) (offset * 100 / FileSize);
+                    int count = (int) ((long)offset * 100 / FileSize);
                     publishProgress(count);
                 }
